Implemented URL query for selected template.
