+++ web-ui/src/components/feedback_template_selector/FeedbackTemplateSelector.test.js
@@ -1,0 +1,8 @@
+import React from "react";
+import FeedbackTemplateSelector from "./FeedbackTemplateSelector";
+import {AppContextProvider} from "../../context/AppContext";
+
+
+it("renders the recipient card", () => {
+  snapshot(<AppContextProvider><FeedbackTemplateSelector/></AppContextProvider>);
+});
