+++ web-ui/src/pages/FeedbackRequestPage.jsx
@@ -1,4 +1,4 @@
-import React, {useState} from "react";
+import React from "react";
 import { makeStyles } from "@material-ui/core/styles";
 import Stepper from "@material-ui/core/Stepper";
 import Step from "@material-ui/core/Step";
@@ -12,6 +12,7 @@ import FeedbackRecipientSelector from "../components/feedback_recipient_selector
 
 import "./FeedbackRequestPage.css";
 import TemplatePreviewModal from "../components/template-preview-modal/TemplatePreviewModal";
+import FeedbackTemplateSelector from "../components/feedback_template_selector/FeedbackTemplateSelector";
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -38,63 +39,21 @@ function getSteps() {
   return ["Select template", "Select recipients", "Set due date"];
 }
 
-function getTemplates() {
-  return [
-    {
-      title: "Ad Hoc",
-      isAdHoc: true,
-      description: "Ask a single question.",
-      creator: "Admin",
-      questions: []
-    },
-    {
-      title: "Survey 1",
-      isAdHoc: false,
-      description: "Make a survey with a few questions",
-      creator: "Admin",
-      questions: []
-    },
-    {
-      title: "Feedback Survey 2",
-      isAdHoc: false,
-      description: "Another type of survey",
-      creator: "Jane Doe",
-      questions: [],
-    },
-    {
-      title: "Custom Template",
-      isAdHoc: false,
-      description: "A very very very very very very very very very very very very very very very very very very very very very very very very very very long description",
-      creator: "Bob Smith",
-      questions: []
-    },
-  ];
-}
-
 const FeedbackRequestPage = () => {
   const steps = getSteps();
   const classes = useStyles();
   const location = useLocation();
-  const query = queryString.parse(location?.search).step?.toString();
-  let activeStep = location?.search ? parseInt(query) : 1;
-  const numbersOnly = /^\d+$/.test(query);
-  const [preview, setPreview] = useState({open: false, selectedTemplate: null});
 
-  const handlePreviewOpen = (event, selectedTemplate) => {
-    setPreview({open: true, selectedTemplate: selectedTemplate});
-  }
+  const query = queryString.parse(location?.search);
 
-  const handlePreviewClose = (selectedTemplate) => {
-    setPreview({open: false, selectedTemplate: selectedTemplate});
-  }
+  const stepQuery = query.step?.toString();
+  const templateQuery = query.template?.toString();
+  // const fromQuery = query.from?.toString();
+  // const sendQuery = query.send?.toString();
+  // const dueQuery = query.due?.toString();
 
-  const onCardClick = (template) => {
-    if (template.isAdHoc) {
-      setPreview({open: true, selectedTemplate: template});
-    } else {
-      console.log(`Selected ${template.title}`);
-    }
-  }
+  let activeStep = location?.search ? parseInt(stepQuery) : 1;
+  const numbersOnly = /^\d+$/.test(stepQuery);
 
   if (activeStep < 1 || activeStep > steps.length || !numbersOnly) {
     return (
@@ -102,23 +61,22 @@ const FeedbackRequestPage = () => {
     );
   }
 
+  const getFeedbackArgs = () => {
+    let args = "";
+    if (templateQuery) args = `${args}&${templateQuery}`;
+
+    return args;
+  }
+
   return (
     <div className="feedback-request-page">
-      {preview.selectedTemplate &&
-        <TemplatePreviewModal
-          template={preview.selectedTemplate}
-          open={preview.open}
-          onClose={handlePreviewClose}
-        />
-      }
-
       <div className="header-container">
         <Typography variant="h4">Feedback Request for <b>John Doe</b></Typography>
         <div>
             <div>
               <Link
                 className={`no-underline-link ${activeStep <= 1 ? 'disabled-link' : ''}`}
-                to={`?step=${activeStep - 1}`}
+                to={`?step=${activeStep - 1}${getFeedbackArgs()}`}
               >
                 <Button
                   disabled={activeStep <= 1}>
@@ -128,7 +86,7 @@ const FeedbackRequestPage = () => {
 
               <Link
                 className={`no-underline-link ${activeStep > getSteps().length ? 'disabled-link no-underline-link' : ''}`}
-                to={activeStep===3 ?`/feedback/request/confirmation` : `?step=${activeStep + 1}`}>
+                to={activeStep===3 ?`/feedback/request/confirmation` : `?step=${activeStep + 1}${getFeedbackArgs()}`}>
                 <Button
                   disabled={activeStep > getSteps().length}
                   variant="contained"
@@ -151,21 +109,7 @@ const FeedbackRequestPage = () => {
         })}
       </Stepper>
       <div className="current-step-content">
-          {activeStep === 1 && <TemplatePreviewModal/> &&
-          <div className="card-container">
-            {getTemplates().map((template) => (
-              <TemplateCard
-                title={template.title}
-                creator={template.creator}
-                description={template.description}
-                isAdHoc={template.isAdHoc}
-                questions={template.questions}
-                expanded={preview.open}
-                onClick={(e) => handlePreviewOpen(e, template)}
-                onCardClick={() => onCardClick(template)}/>
-            ))}
-          </div>
-        }
+        {activeStep === 1 && <FeedbackTemplateSelector/> }
         {activeStep === 2 && <FeedbackRecipientSelector />}
       </div>
     </div>
